backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
PyJWT>=2.8.0
reportlab>=4.0
boto3>=1.28.0
